removed css files and added google logic to verify/reset
