--- app/src/AccountRecovery.css
@@ -1,8 +1,0 @@
-.account-recovery-box {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.account-recovery-box .MuiAvatar-root {
-  margin-top: 20px;
-}
