--- app/src/AccountVerify.css
@@ -1,5 +1,0 @@
-.account-verify-box {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
