+++ server/src/util.ts
@@ -2,9 +2,6 @@ import fs from "fs";
 import path from "path";
 import fsPromises, { rm } from "fs/promises";
 import { glob } from "glob";
-import nodemailer from "nodemailer";
-import dotenv from "dotenv";
-import { Email, assertEmail } from "./types.js";
 
 /**
  * Recursively deletes directory INCLUDING directoryPath
@@ -158,30 +155,3 @@ export async function resetUploadDirectory(uploadedFile: string, decryptedUpload
     deleteDirectory(decryptedUploadedDirectory);
   }
 }
-
-dotenv.config();
-const transporter = nodemailer.createTransport({
-  service: "gmail",
-  host: "smtp.gmail.com",
-  port: 587,
-  secure: false,
-  auth: {
-    user: process.env.EMAIL_USER,
-    pass: process.env.EMAIL_PASS
-  }
-});
-
-export const sendEmail = async (email: Email) => {
-  assertEmail(email);
-  try {
-    await transporter.sendMail({
-      from: email.from,
-      to: email.to,
-      subject: email.subject,
-      text: email.text
-    });
-  } catch (error) {
-    console.error("An error occurred:", error);
-    throw error;
-  }
-};
